refactor: refactor the actor component and update the loading UI for the actor slide
