+++ components/actor/actor-item.tsx
