+++ components/movie/movie-content.tsx
@@ -1,7 +1,7 @@
 import Country from 'types/country';
 import DetailMovie from 'types/detail-movie';
 import Trailer from './trailer';
-import ActorList from '../commons/actor-list';
+import ActorList from '../actor/actor-list';
 import Credit from 'types/credit';
 import isNonEmpty from 'utils/is-none-empty';
 import MovieSummary from './movie-summary';
